refactor: update imports for AlertPropsOptionalKey to useAlerts hook and adjust EstimatedCostInfo import
