+++ src/hooks/useChallengeWebSocket.tsx
@@ -1,8 +1,8 @@
 import { SS58String } from 'polkadot-api';
 import { useEffect, useCallback, useState, useRef } from 'react';
-import { AlertPropsOptionalKey } from '~/store/AlertStore';
 import { ChallengeStatus, ChallengeStore } from '~/store/challengesStore';
 import { IdentityInfo, verifyStatuses } from '~/types/Identity';
+import { AlertPropsOptionalKey } from './useAlerts';
 
 // Types matching your Rust backend
 type Data = {
